Move consts from node_consts.py to actor_consts.py
