+++ dashboard/modules/actor/actor_consts.py
@@ -1,0 +1,5 @@
+import ray
+
+ACTOR_CHANNEL = "ACTOR"
+NIL_NODE_ID = ray.NodeID.nil().hex()
+RETRY_GET_ALL_ACTOR_INFO_INTERVAL_SECONDS = 1
