+++ dashboard/modules/node/node_consts.py
@@ -1,9 +1,4 @@
-import ray
-
 NODE_STATS_UPDATE_INTERVAL_SECONDS = 1
-RETRY_GET_ALL_ACTOR_INFO_INTERVAL_SECONDS = 1
-ACTOR_CHANNEL = "ACTOR"
 ERROR_INFO_UPDATE_INTERVAL_SECONDS = 5
 LOG_INFO_UPDATE_INTERVAL_SECONDS = 5
-NIL_NODE_ID = ray.NodeID.nil().hex()
 UPDATE_NODES_INTERVAL_SECONDS = 5
